Update Makefile
Sklaffkom has logging facilities already built in but the /log directory was not created during install. I consider this a bug in the makefile, and here's a quick and dirty fix.

To enable debug logging, set LOGLEVEL in sklaff.h to a high number, I set it to 100 but 30 should be enough. 

To make it even more verbose just insert a line such as this one wherever needed (the number reflects the loglevel) :

debuglog("user already logged in", 7);